add: endpoint and controller for assign role to a user

+++ app/Http/Controllers/UserController.php
@@ -5,17 +5,27 @@ namespace App\Http\Controllers;
 use App\Http\Requests\RegisterFormRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\Rules\Password;
+use Spatie\Permission\Models\Role;
 
 class UserController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
         $users = User::all();
+
+        if ($request->has('search')) {
+            $users = User::where('name', 'like', '%' . $request->search . '%')
+                ->orWhere('email', 'like', '%' . $request->search . '%')
+                ->get();
+        }
 
         return response()->json(UserResource::collection($users), 200);
     }
@@ -25,11 +35,7 @@ class UserController extends Controller
      */
     public function store(RegisterFormRequest $request)
     {
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-        ]);
+        $user = User::create($request->validated());
         $user->email_verified_at = now();
         $user->save();
 
@@ -54,9 +60,18 @@ class UserController extends Controller
         $validated = $request->validate([
             'name' => ['sometimes', 'string', 'max:255'],
             'email' => ['sometimes', 'string', 'email', 'max:255'],
+            'password' => ['sometimes', 'string', 'confirmed', Password::defaults()],
         ]);
 
-        $user->update($validated);
+        $user->update([
+            'name' => $validated['name'] ?? $user->name,
+            'email' => $validated['email'] ?? $user->email
+        ]);
+
+        if (isset($validated['password'])) {
+            $user->password = Hash::make($validated['password']);
+            $user->save();
+        }
 
         return response()->json(new UserResource($user), 200);
     }
@@ -66,6 +81,31 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if ($user->id === Auth::user()->id) {
+            return response()->json(['message' => 'No puedes eliminar tu propio usuario, imbecil.'], 403);
+        }
+
+        if ($user->hasRole('admin')) {
+            return response()->json(['message' => 'No puedes eliminar a un administrador.'], 403);
+        }
+
         $user->delete();
+
+        return response()->json([
+            'message' => 'Usuario eliminado correctamente.',
+            'user' => new UserResource($user)
+        ], 200);
+    }
+
+
+    public function assignRoleToUser(User $user, string $role): JsonResponse
+    {
+        if (!Role::where('name', $role)->exists()) {
+            return response()->json(['error' => 'El rol no existe.', 'roles' => Role::all()->pluck('name')], 404);
+        }
+
+        $user->assignRole($role);
+
+        return response()->json(['message' => 'Rol asignado correctamente.', new UserResource($user)], 200);
     }
 }
